Restructure the components and fix some warnings
